Added authentication and security measures to login page
Incorporated event handlers to ensure multiple factors were in place when user is logging in. For one, the user's email must end in .edu to ensure that it is a college student logging in. There are also other security measures in place like a minimum 6 character password, checks on email formatting, and checks on non-null values for email and password.