Added sign up functionality to app
Previously only login in functionality was available on app, so new users didn't have a way of joining our app. Now, on the login page, there is a place that prompts new users to Sign Up. This button takes them to a new Sign Up page to create their account and successfully join the app. The same security and authentication issues are addressed with Sign Up: emails must end in .edu to ensure only college students use our app, as well as a password length minimum (6) and checks on email formatting.